+++ package.json
@@ -6,15 +6,15 @@
     "bench": "npm -C bench run bench",
     "build": "esno scripts/copy-files.ts && pnpm -r --filter !playground run build && pnpm -r run build-post",
     "dev": "nr stub",
-    "stub": "pnpm -r --parallel run stub",
+    "docs": "npm -C docs run dev",
+    "docs:build": "npm -C docs run build",
     "lint": "eslint \"**/*.{ts,vue,mjs}\"",
     "lint:fix": "nr lint -- --fix",
     "play": "npm -C playground run dev",
     "play:build": "pnpm run build && npm -C playground run build",
     "release": "bumpp package.json packages/*/package.json --commit --push --tag",
     "size": "esmo scripts/size.ts",
-    "docs": "npm -C docs run dev",
-    "docs:build": "npm -C docs run build",
+    "stub": "pnpm -r --parallel run stub",
     "test": "vitest test",
     "test:update": "vitest -u"
   },
@@ -38,7 +38,6 @@
     "@unocss/cli": "workspace:*",
     "@unocss/config": "workspace:*",
     "@unocss/core": "workspace:*",
-    "@unocss/css-transform": "workspace:*",
     "@unocss/extractor-pug": "workspace:*",
     "@unocss/preset-attributify": "workspace:*",
     "@unocss/preset-icons": "workspace:*",
@@ -48,9 +47,10 @@
     "@unocss/preset-web-fonts": "workspace:*",
     "@unocss/preset-wind": "workspace:*",
     "@unocss/reset": "workspace:*",
+    "@unocss/transformer-css-directive": "workspace:*",
+    "@unocss/transformer-variant-group": "workspace:*",
     "@unocss/vite": "workspace:*",
     "@unocss/webpack": "workspace:*",
-    "@unocss/transformer-variant-group": "workspace:*",
     "@vitejs/plugin-vue": "^2.1.0",
     "@vitest/ui": "^0.2.7",
     "@vueuse/core": "^7.5.5",
