--- packages/css-transform/README.md
@@ -1,7 +1,0 @@
-# @unocss/css-transform
-
-CSS Transformation for UnoCSS
-
-## License
-
-MIT License © 2021-PRESENT [Anthony Fu](https://github.com/antfu)
