+++ packages/transformer-css-directive/build.config.ts
@@ -8,7 +8,6 @@ export default defineBuildConfig({
   clean: true,
   declaration: true,
   externals: [
-    'pug',
     '@unocss/core',
   ],
   rollup: {
