+++ packages/transformer-css-directive/package.json
@@ -1,11 +1,12 @@
 {
-  "name": "@unocss/css-transform",
+  "name": "@unocss/transformer-css-directive",
   "version": "0.25.0",
-  "description": "CSS Transformation",
+  "description": "UnoCSS transformer for `@apply` directive",
   "keywords": [
-    "unocss"
+    "unocss",
+    "unocss-transformer"
   ],
-  "homepage": "https://github.com/antfu/unocss/tree/main/packages/css-transform#readme",
+  "homepage": "https://github.com/antfu/unocss/tree/main/packages/transformer-css-directive#readme",
   "bugs": {
     "url": "https://github.com/antfu/unocss/issues"
   },
@@ -13,7 +14,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/antfu/unocss.git",
-    "directory": "packages/css-transform"
+    "directory": "packages/transformer-css-directive"
   },
   "funding": "https://github.com/sponsors/antfu",
   "author": "Anthony Fu <anthonyfu117@hotmail.com>",
