+++ packages/vite/src/index.ts
@@ -10,7 +10,6 @@ import { SvelteScopedPlugin } from './modes/svelte-scoped'
 import { ShadowDomModuleModePlugin } from './modes/shadow-dom'
 import { ConfigHMRPlugin } from './config-hmr'
 import type { VitePluginConfig } from './types'
-import { transformCSSPlugin } from './transform/css'
 import { initTransformerPlugins } from './transformers'
 
 export * from './types'
@@ -19,8 +18,6 @@ export * from './modes/global'
 export * from './modes/per-module'
 export * from './modes/vue-scoped'
 export * from './modes/svelte-scoped'
-
-export type { UnocssPluginContext } from '../../plugins-common'
 
 export function defineConfig<Theme extends {}>(config: VitePluginConfig<Theme>) {
   return config
@@ -33,7 +30,6 @@ export default function UnocssPlugin(
   const ctx = createContext<VitePluginConfig>(configOrPath, defaults)
   const inlineConfig = (configOrPath && typeof configOrPath !== 'string') ? configOrPath : {}
   const mode = inlineConfig.mode ?? 'global'
-  const transformCSS = inlineConfig.transformCSS ?? false
 
   const plugins = [
     ...initTransformerPlugins(ctx),
@@ -66,15 +62,6 @@ export default function UnocssPlugin(
   }
   else {
     throw new Error(`[unocss] unknown mode "${mode}"`)
-  }
-
-  // CSS transform
-  if (transformCSS) {
-    const cssPlugin = transformCSSPlugin(ctx)
-    if (typeof transformCSS === 'string')
-      cssPlugin.enforce = transformCSS
-
-    plugins.push(cssPlugin)
   }
 
   return plugins.filter(Boolean) as Plugin[]
