+++ packages/vite/src/modes/global/index.ts
@@ -1,4 +1,4 @@
-import type { UnocssPluginContext } from '../../../../plugins-common'
+import type { UnocssPluginContext } from '@unocss/core'
 import { GlobalModeBuildPlugin } from './build'
 import { GlobalModeDevPlugin } from './dev'
