+++ playground/src/logics/config.ts
@@ -1,3 +1,4 @@
+import type { UserConfig } from 'unocss'
 import { createConfig } from '../../unocss.config'
 import { options } from './url'
 
@@ -5,5 +6,5 @@ export const defaultConfig = computed(() =>
   createConfig({
     strict: !!options.value.strict,
     dev: true,
-  }),
+  }) as UserConfig<any>,
 )
