+++ playground/unocss.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig, presetAttributify, presetUno } from 'unocss'
 import presetIcons from '@unocss/preset-icons'
 import transformerVariantGroup from '@unocss/transformer-variant-group'
+import transformerCSSDirective from '@unocss/transformer-css-directive'
 
 export function createConfig({ strict = true, dev = true } = {}) {
   return defineConfig({
@@ -28,6 +29,7 @@ export function createConfig({ strict = true, dev = true } = {}) {
     ],
     transformers: [
       transformerVariantGroup(),
+      transformerCSSDirective(),
     ],
   })
 }
